Add thoughts on operators
R has three levels of feedback builtin: errors, warnings, and messages.
These correspond quite nicely with the three levels I have in the monad.
I think rather than using `warn`, `fail`, and `note`, I can just use
function that call the R builtins, e.g `warning`, `stop`, and `message`.

The monad handles the effects R produces when running functions.
Previously I felt sore about having a monad library that offered only
one monad. However, thinking about it in terms of a monad that handles R
effects, then this is not just A monad, it is THE monad. The Rmonad. So
I have renamed the repository. From monadR, a general monad class, to
Rmonad, a specific monad. This is also a nod to Xmonad.