Refactor as tree, and mfind
This is mostly an internal implementation change. I converted from a
representation that distinguished between the top monad and the list of
historical records. Now present and past are equal, all are Rmonad
objects. The present data have links to their parents, but not vise
versa. Eventually I should refactor this again using a formal graph
structure, maybe something from igraph.

Also I added an mfind function, which allows specific historical
computations to be extracted.

This whole new system is not well tested yet. I need to add a suite of
tests.