Partial solution to nest relinking
How should nesting of pipelines in functions be handled?

The current implementation simply links the output of the nested
function, the final monadic container, to the input. This essentially
collapses all the interior nodes. While this yields correct ouput,
history is lost. Referential transparency is lost as well. A named
function should be able to be swapped for a section of the pipeline
without altering the results.

My approach to solving this problem is, perhaps, a little drastic.
Before execution, I can disembowel the function. All variables local to
the function are declared before the first monadic operator (unless the
programmer uses `assign` or `<<-`). They can be extracted. The internal
pipeline can be then be laid out, the globs that use bound variables,
can be linked, etc.

Anyway, so far I have a set of functions for dissecting expressions and
the tests needed to keep the functions in line.