Complete rough nested pipeline relinking
The approach now is to build a matrix linking local variables to
dependencies on the monadic inputs. Then the code contained in each
nested node can be checked for free variables that depend on the monadic
inputs. Nodes are relinked to their dependencies (parents) as needed.

The raw code is mostly together. It doesn't work yet, though, and hasn't
been connected to bind.

In order for this to work, I need to make '%*>%' and friends store the
code in its final state (after adding arguments to naked bodies and the
like) rather than initial state (the raw little thing the programmer
gave rmonad).