Add plot to gff-processing vignette
I had to use a bit of a hack to get this working. When I try to plot
using knitr, with:

```{r, results='asis'}
plot(result)
```

I get 5Mb of scripts prepended to my HTML file. Also the HTML file
freezes on loading, with a message saying the scripts are not
responding.

I think this is a problem with the particular vignette engine I am
using, since

```
$ Rscript -e "library(knitr); knit('filename.Rmd')"
```

works. The difference is that the script above makes a PNG file for the
plot and imports it into the resulting Markdown file. Whereas the
vignette builder includes the DOT code for the graph, all the scripts
needed to build the DOT code, and calls them at runtime.