Fix linking of branched nodes
Now for an operation such as `x %>^% foo %>>% bar`, `foo` and `bar` will
both be linked as the children of `x` with now relation to eachother.
Also `foo` may be a nested expression.

There is a lingering issue of the phantom node. I found that for the
above function, `as.list` yields 4 nodes, rather than 3. This is because
the branch operator (`%^>%`) passes the lhs monad to the right, then the
`%>>%` bind operation adds `bar` as a child of `x`. Since the S4 class
is immutable, this creates a copy. I may need to convert to R6 classes,
which use reference semantics.