Borken refactor into R6
The code is almost right. A bunch of test fail for reasons I will figure
out tomorrow.

My refactoring approach is to leave almost all of my original code
intact, changing only the implementation of the Rmonad class. So I have
getters and setters in the class itself, and all my old accessors are
wrappers for these getters and setters. It *may* be simpler to remove
all this extra code and just use the R6 methods. However, it is often
useful to pass accessors as arguments (e.g. into lapply).

A few things to do:

 * replace the `branch` field with `children`

 * add links from parent to child (only child to parent made currently).

 * write recursion functions into the class

 * replace the `%>^%` operator with a branching equivalent of `funnel`