Having serious misgivings about the R6 classes
My old approach was problematic since it duplicated nodes that ought to
have singular identities. The R6 approach is problematic because it
doesn't.

There is a good reason reference semantics are avoided in functional
programming. R is incomprehensible enough as it is. Already in `rmonad`
I have to deal with NSE and weird environmental dependencies. I really
don't want to dabble in cloning as well.

What to do? Perhaps it is time I begin building that ultimate data
structure I've always wanted -- using a real graph library.

In this context, `as_monad` would produce a singleton graph (unless
there was nesting). Bind operations would add new elements to the graph.
Data would be stored in a list indexed with node ids. The values are
deleted with when the node is no longer on the "surface" of the graph.
Since we are not dealing in reference semantics, deletion of the value
will not magically be mirrored in nodes defined elsewhere.

Nests are a little complex, but I have already developed an approach for
dealing with them (through edge annotations).

I would probably need to supplement the graph with extra data (e.g. the
ids of the "head" node).