Resolve check issues
All seems to be working well, too well. I am not happy with the R6
packages, but I did resolve all the immediate problems. I will attempt
to run my Fagin pipeline with the new rmonad version, if this works, I
will begin the CRAN submission process for v0.2.0. This will require
some additional documentation and a careful review of the README and
vignettes.