Add test for issue #2
The `%__%` operator associates two pipelines. No variables from the left
hand side can be used on the right, but the left may have altered state
(e.g. created files), so this is still a directed edge. However, it is
quite unlike the parent-child, nested, or branching relationships I
currently handle.