Encode prior pipelines as new kind of edge
Now there is a new field ('prior') in the Rmonad object that holds a
link to a previous pipeline. For example,

```
session_info() %__% iris %>>% head
```

`session_info` is not the parent of the subsequent code, but needs to be
associated with it. Whatever is left of the `%__%` will not pass values
to the right (though it may affect state), so represents the sole link
between two components of the graph.

Importantly, `%__%` now stores the final result of the left hand
computation. So the `%v__%` operator is no longer needed.

TODO:

 * deprecate `%v__%`
 * add dedicated plot and print handling
 * update the vignettes and README