Fix issue #3
This still feels a little hacky. Now the final result is stored at the
place where there error occurs, but the higher nest containers that
inherit from it are set to NA. It might be more natural if each of them
contained the same final result. Or better if each linked to the last
valid result at its level. Ah, this is the way to do it.