Reimplement most of the accessors
I still need to work out the igraph link setters (`m_parents<-`,
`m_branch<-`, etc). I did write the getters, though I am not confident
that I did it the best way. The getters return the Rmonad object with
the head pointing to the requested nodes (e.g. the parents) or to NULL
on failure (e.g. parent of root). The problem with this is that the
vertex ids are not stable.

There are a couple other little issues I've run into; these are
described in scattered TODO and FIXME statements. I'll address them once
I have the gist of the refactor working.

The next thing I need to do is write the link setters. Then I can work
on getting a functional `as_monad` function. After `as_monad`, I can
rewrite `as.list` and then fix the operators and bind. Then I can go on
to the dreaded rewrite of magic. After all this, my tests should pass
again.

After the tests pass, I will need to rewrite the print and plot
functions and update the vignettes, README, and other docs. I'll also
need to add new tests, as appropriate.

I'll stop after this, do some cleanup, and submit to CRAN -- leaving the
next stage of building upon the igraph core for v0.5.