Remove the `recursive_set_nest_depth` function
That function was an abomination. I previously did not set nest depth
until needed, leaving the values incorrect in the meantime. Nest depth
should be set when the nest graph is linked to downstream high child. I
just have to recursively increment the depth following all edges except
"transitive" (or whatever it is I am calling the edges that connect a
nested node to a higher input).