Remove as_dgr_graph and stubify Rmonad.plot
Now that the Rmonad objects are igraph based, there isn't really any
need for a special function to convert to DiagrameR. I may add this
function back at some point. It does make sense to convert from the rich
Rmonad object to a specialized form for easy plotting.

The plotting function I obviously will keep. But I will have to make
huge changes to it, and will be plotting in igraph, so not much of the
current code can be reused. It is cleaner to start ab initio.