Add vectorized accessors
For now I am using `m_*` to get the head attribute and `ms_*` to get the
vetorized output. I don't like th convention. But it works for now.
Importantly, it is backwards compatible (though already this coming
version is backwards incompatible).

I have tests for warning, note, and error accessors. This tests the
concept, but I need coverage of the other values, especially `value`,
since that one is so tricky.