Partially fix `%>^%`; vectorized link functions
I've added `ms_branch`, `ms_parents`, `ms_nest`, and `ms_prior`
functions. These are all lists of integer vectors, holding either NULL,
if there are not links, or a list of ids.

Branching still not work quite like it should, but I've fixed the
obvious bugs.

One issue is that an extra node is being created by `%>^%`, e.g.

```
16 %>^% sqrt %>% ms_value
```

This should report `[16, 4]`, but instead it reports `[16,4,16]`.