Make indexed accessors and resurrect print
Now every non-vectorized getter takes an index argument. This allows any
particular value to be directly accessed. Note the setters still only
work on the head node. To allow the setters to work, I would need to
write a generic `[` function. Though this is probably a reasonable thing
to do anyway.