Add indexing to `ms_*` accessors with tests
Now `m_*` accessors are explicitly single value accessors and `ms_*` are
explicitly multi-value accessors (return things in lists, even if there
is only one element). The both accessors take an optional `index`
arguemtns. In the case of `m_*`, the argument must be of length 1 and
defaults to `head`. For `ms_*`, the arguement can be any length and
defaults to every element.

I also add tests for subsetting with `ms_*`.