Vectorize `has_*` functions
These functions pass arguments to the `ms_*` functions, which include
`index`. So to get just the head result: `has_*(x, index=x@head)`.
Externally, the most common use of the `has_*` functions is probably in
subsetting vectors (as is done in the mtabulate function). Internally
there of handling of just the head as the graphs are built. But there is
not so much use as to justify adding a singular set of functions. Also,
some use of the singular, such as in `print`, I think could be better
replaced with the vectorized version.