Replace accessor prefix `ms_` with `get_`
Normally in R getters and setters share names. However, rmonad objects
are written internally and accessed externally. There is normally no
need for the user to set values inside the object. The `get_` prefix
makes this very clear. Certainly it is better than than the absurd
`ms_`, which seems to be referring to the functions as unmarried ladies.