A little closer - changed summary to list
Many tests still do not pass. Transitive links are broken. Something is
wrong with nesting.

In this commit I also changed `summary` to a list, rather than singular
value. The rewriter now accepts a list of functions. This will allow
summary to generalize to a list of results that will be included in the
report (I need to update report to handle this).