Bump dev version to 0.4.0.9004
Rmonad is back to a passing state. Jubilation. There are a few things
that need to be done, though, before it is CRAN-ready.

 1. factor out the `*raw_value*` functions

 2. update report to work with the new list-version of summary

 3. update plot to handle time and space coloring

There are lots of big plans I haven't gotten around to implementing yet:

 1. global/graph settings to allow dry-runs, setting default cache,
    adding run environments, changing verbosity levels, etc.

 2. resume functions (making the caches actually useful)

 3. full code capture (including mapping the operators into the edges)

 4. Rmonad diff

 5. some sort of system for typechecking inputs and outputs

 6. node factories (and inheritance)

 7. nest attributes (e.g. allow distinction between parallel nodes in a
    loop). Perhaps allow packaging of looped things, such that each node
    is a vector across parallel states. Following the
    split-compute-recombine strategy.

Anyway, there are lots of things to do. But these can all wait for a
future version.

For v0.4, I want a clean, minimal, igraph-based implementation, which
can serve as the foundation for the fancy features to come.