Remove the indexing on the setters
Umm, yeah, so there was a reason I didn't pass `...` to my setters,
because they were, of course, binary operators. So the `xxx<-` setters
can't be indexed. This does, however, leave me with no direct way to
alter the interior nodes of an Rmonad. Now I don't usually need to alter
the interior, so maybe this is fine?