Rename cache functions and add preserve option
Probably the main use of the recacher is to take a transient
intermediate value and cache (probably to a file). Previously these
functions only changed the cache paradigm, but left the 'store' field
unchanged, which meant the value would still be deleted on the next
bind. Now, by default, the recachers set 'store=TRUE'.