Add Rprof.out to gitignore
I've found Rmonad v0.3.0 was 20X faster, but still hit the recursion
ceiling when trying to run 1000 sequential `%>>%` operations.

There is something deeply wrong about the way I am parsing the
pipelines. The lhs input to bind is generally a long expression (e.g.
for `a%>>%f%>>%g%>>%h`, it will be `a%>>%f%>>%g`). I evaluate this
recursively, which is not efficient. Compare this to the `magrittr`
implementation, which splits the pipeline into a list and iterates over
it.

However, nothing about the core binding implementation changed in v0.4
(apart from splicing, but that wouldn't affect this benchmark). So in
the 20X decrease is probably due to the igraph change.

Specifically, the simplest use of `as_monad` is much slower now: 0.7 ms
in v0.3 versus 7 ms now.