Remove Rmonad(), use as_monad(), instead
Previously Rmonad() created a valueless Rmonad object. Such an object is
never used directly and overlaps enormously with `as_monad`. So, for
now, I join them. I am not yet happy with the situation, though.

This simplification also speeds up as_monad by around 40%. However, this
speedup only helps when the input is not an Rmonad. It also does nothing
to speed chains of `%>>%`.

No, the real culprit is that abomination `.rmonad_union`. The unnamed
unions, where no attribute mangling occurs, are fast (200 us). But with
all the attribute trash, it bogs down to 6ms. One solution would be to
extract all the attributes into a named list stored in the Rmonad
object. Then the underlying graph would be attribute free. Or I could
use `triebeard`, or some such program to store the values, that would
avoid the costs or reallocating the list on every bind.