Remove performance tests
I've posted two github issues about 1) slow performance and 2) the crash
on long expressions.

Benchmarking is not something that is very easy to test automatically
(being very system specific).

The second problem is not likely to be encountered in practice. The
fagin pipeline, which involves thousands of nodes, runs without problem.
Only sequential runs of hundreds of functions crashes. This is probably
not something people will often do.