A first crack at writing the NEWS
There is a lot of work to be done here, I pretty much trashed everything
and rewrote it. No tidy deprecation and backwards compatibility here, I
nuked everything.

Well, that is an overstatement. The exported functions, the API, is
basically the same (with some renaming and argument changes). But a lot
of little details have changed (you know, those little details in R that
always poke you in the eye).