Do not export `bind`
There really is no reason why any user should ever use `bind` directly.
Furthermore, the function is not stable. It is the core of rmonad, but
it is also a hot mess and a major target for heavy refactoring. I want
the freedom to totally rewite the function without breaking downstream
code.

So, no export for bind.