Generalize make_local_cacher
Now the `make_local_cacher` takes arguments for specializing the method
of saving, retrieving, deleting, and checking the cache. The one thing
that is not generalized is the filename handling. I am inclined to leave
this inaccessible, since the user shouldn't have to bother with
filenames. However, the users, may disagree. If they do, they are
welcome to argue their case in the github issues.