First draft of loop function
I do not like this. I need a real solution, a formalism to base all this
mess on. But who has time for math?

The new `loop` function takes an Rmonad that wraps a vector, and applies
a function across the elements. Each function returns an Rmonad object.
`loop` merges these objects into one (with `combine`) and links the
resulting node to the input. This preserves history, by linking the
wrapper around all the runs to the parent holding the input vector, but
no link is created between each run and its specific input. The user can
provide this link through the tag system, but I would prefer it were
automated.