Spiral further into madness
I've sorted out key handling mostly. My approach of XORing together the
parents is not quite right. XOR is independent of order but pipeline
should be. This means that pipelines with the same nodes but different
orders will have the same final key. I could add pipeline depth to the
key, in some manner, though this is currently not something I track.

The package is still pretty far from running.