Add nest salt
Dealing with nests is really hard. The nests are dependent on their
parnents. But the nested function does not "know" they are a nest, I
can't directly pass anything to them. So I am passing information
through their scope, dynamic scope. I set a value in the enclosing
environment of the function that initializes the nest. Then in as_monad
and bind I dynamically search through the call stack for this value, and
append it to the serialized code before the digest. Each time I set the
nest_salt, I append it to any existing salt.

This is a really scary solution. But I am not sure of a better way.

Now `loop` works. and 7/354 tests fail.