Clean up digests, remove XOR statements
I really should not rely on XOR anywhere in a DAG. There is a distinct
possibility of XORing against the same value twice, which cancels out
the dependency. It is much safer to just concatenate the serialized
values of everything I want a key to depend on and then take the hash.