Dev release v0.5.0.9009
The new caching system is roughly working. I am not quite happy with it.
the automatic caching is nice, simplifies things certainly. But it can
also cache when you don't really want it to. Also the performance cost
is pretty high.

I ought to add a way to store the value in memory AND cache it for
later.

TODO:

 - [ ] Write the `crunch` function

 - [ ] Write the `archive` function