Add test that breaks metadata handling
The following statement fails:

```
foo <- function(x) { list(x=x); x }
5 %>>% foo
```

The problem is that when the metadata list is evaluated, `x` is not in
scope. It is evaluated in the parent frame, rather then the functional
enclosure itself. The RIGHT way to do it is to evaluate the list in the
scope of the arguments, which inherits immediately from the parent
frame. That is, the `list(x=x)` should be exactly what the programmer
would expect it to be at that location in the function.