Fix issue with identical nests
To do this, I add the key within each bind to the salt. This means any
nest that occurs in the evaluated expression will depend on the parent
key. Which is what I want, I suppose. It isn't perfect and will result
in some things being unnecessarily re-evaluated. For example:

```R
 # Here I cache the "expensive_operation(45)" expression result
 foo <- as_monad(expensive_operation(45))

 # Here I should be able to reuse the result, but I can't, because the
 # key is dependent on the parent (the node holding "5").
 5 %>>% { expensive_operation(45) %>>% prod(.) }
```