(4 fail) Partial refactor with muliple tags
A node needs to be able to have multiple tags. This is needed especially
for adding tags to failed nodes. At first sight, this may seem like a
bad idea, but failed nodes may need to be accessed along the chain. And
if no tag is made, errors will be raised outside rmonad when the missing
tag is accessed (via 'view'), leading to loss in history.

I've implemented most of the handling. But there are still some tests
that don't pass.